// src/app/models/pago.modelo.ts
export class Pago {
    id?: number;
    card_number: string;
    exp_year: number;
    exp_month: number;
    cvc: number;
    name: string;
    last_name: string;
    bill: string;
    value: number;
    phone: string;
    doc_number: string;
    city: string;
    address: string;
    cell_phone: string;
    email: string;

}